# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ArmatureInfo": {
        "additionalProperties": false,
        "properties": {
          "bones": {
            "items": {
              "$ref": "#/components/schemas/BoneInfo"
            },
            "title": "Bones",
            "type": "array"
          },
          "name": {
            "maxLength": 63,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.\\-: ]+$",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "bones"
        ],
        "title": "ArmatureInfo",
        "type": "object"
      },
...
      "Body_upload_scene_v1_scenes_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_scene_v1_scenes_post",
        "type": "object"
      },
      "BoneInfo": {
        "additionalProperties": false,
        "properties": {
          "deform": {
            "default": true,
            "title": "Deform",
            "type": "boolean"
          },
          "name": {
            "maxLength": 63,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.\\-: ]+$",
            "title": "Name",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "maxLength": 63,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.\\-: ]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          }
        },
        "required": [
          "name"
        ],
        "title": "BoneInfo",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "SCENE_NOT_FOUND",
          "JOB_NOT_FOUND",
          "CANDIDATE_NOT_FOUND",
          "ARTIFACT_NOT_FOUND",
          "ARMATURE_NOT_FOUND",
          "BONE_NOT_FOUND",
          "INVALID_FRAME_RANGE",
          "UNSUPPORTED_REPAIR_TYPE",
          "UNSUPPORTED_RIG",
          "INVALID_STATE",
          "NOT_READY",
          "IDEMPOTENCY_CONFLICT",
          "FILE_INVALID",
          "FILE_TOO_LARGE",
          "VALIDATION_ERROR",
          "NOT_IMPLEMENTED",
          "PLAN_INVALID",
          "PROVIDER_UNAVAILABLE",
          "PROVIDER_TIMEOUT",
          "PROVIDER_RATE_LIMITED",
          "WORKER_CRASHED",
          "WORKER_TIMEOUT",
          "WORKER_OUTPUT_INVALID",
          "RENDER_FAILED",
          "NO_VIABLE_CANDIDATE",
          "STORAGE_ERROR",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "An RFC 7807 problem detail, extended with a machine-readable ``code``.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SceneRef": {
        "additionalProperties": false,
        "description": "An uploaded scene, inspected on upload so that job requests can be validated.",
        "properties": {
          "armatures": {
            "items": {
              "$ref": "#/components/schemas/ArmatureInfo"
            },
            "title": "Armatures",
            "type": "array"
          },
          "blender_version": {
            "maxLength": 32,
            "title": "Blender Version",
            "type": "string"
          },
          "fps": {
            "exclusiveMinimum": 0.0,
            "title": "Fps",
            "type": "number"
          },
          "frame_end": {
            "title": "Frame End",
            "type": "integer"
          },
          "frame_start": {
            "title": "Frame Start",
            "type": "integer"
          },
          "scene_id": {
            "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
            "title": "Scene Id",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "scene_id",
          "sha256",
          "size_bytes",
          "frame_start",
          "frame_end",
          "fps",
          "blender_version",
          "armatures"
        ],
        "title": "SceneRef",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI plans and evaluates; deterministic code executes; the animator decides. Every error body is an RFC 7807 Problem carrying a machine-readable `code`.",
    "summary": "Localized animation repair: measure, propose, compare, apply.",
    "title": "Kinesis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/scenes": {
      "post": {
        "description": "Accepts a single `.blend` file (at most KINESIS_MAX_UPLOAD_MB; it must start with the `BLENDER` magic bytes). The client filename is ignored. The file is stored as `input/scene.blend` in a new scene directory and inspected headlessly with auto-run scripts disabled.",
        "operationId": "upload_scene_v1_scenes_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_scene_v1_scenes_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Upload a .blend copy and inspect it",
        "tags": [
          "scenes"
        ]
      }
    },
    "/v1/scenes/{scene_id}": {
      "get": {
        "operationId": "get_scene_v1_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{5,63}$",
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Get an uploaded scene's inventory",
        "tags": [
          "scenes"
        ]
      }
    },
...
  }
}